chore: refactor nuget parser
