+++ package.json
@@ -25,6 +25,7 @@
     "debug": "^3.1.0",
     "jszip": "^3.1.5",
     "lodash": "^4.17.10",
+    "snyk-paket-parser": "^1.2.0",
     "xml2js": "^0.4.17"
   },
   "devDependencies": {
